refactor: 💡 change product validations
changed quantity and price validations to support only integer values
