+++ app/Http/Requests/UpdateProductRequest.php
@@ -17,8 +17,8 @@ class UpdateProductRequest extends FormRequest
             'name' => ['required', 'min:3', 'max:100'],
             'images.*' => ['image', 'mimes:jpeg,png,jpg,gif,svg', 'max:2048'],
             'category_id' => ['required'],
-            'price' => ['required', 'numeric', 'min:1'],
-            'quantity' => ['required', 'numeric', 'min:0'],
+            'price' => ['required', 'integer', 'min:1'],
+            'quantity' => ['required', 'integer', 'min:0'],
             'description' => ['required', 'min:10', 'max:250'],
             'status' => ['required', 'in:New, Used'],
         ];
